feat(237-03): extract semantic search helpers
- move actor enrichment, vector rank lookup, and RRF merge helpers into service_semantic.py

- keep _compute_rrf_scores available from the public search service facade

- align hybrid tests with CatalogPort-based semantic dispatch
